Драйвер SPI готов

+++ include/driver/spi.h
@@ -6,4 +6,4 @@
 /* Prototypes */
 
 SPI_Status_t SPI_SendData(uint8_t *data, uint32_t size);
-SPI_Status_t SPI_ReadData(uint16_t *data, uint32_t size);
+SPI_Status_t SPI_ReadData(uint8_t *data, uint32_t size);
